# Optimize memory usage of impute_accessibility and normalize_scores