Optimize memory usage of impute_accessibility and normalize_scores:
 
```
                           Wall time        Peak memory     Memory increment
Old version
    impute_accessibility    5min 48s        216295.86 MiB   210960.24 MiB
    normalize_scores        3min 27s        267570.90 MiB   213819.33 MiB
New version
    impute_accessibility    49.6 s          62451.41 MiB    57011.63 MiB
    normalize_scores        2min 16s        160660.14 MiB   106916.97 MiB
Fraction improvement
    impute_accessibility    7x              3.5x            3.7x
    normalize_scores        1.5x            1.6x            2.0x
```