@SeppeDeWinter 2023-04-28
Tested using a matrix of size (regions x cells) 488_211 x 28_705


# Summary

```
                            Wall time        Peak memory     Memory increment
Old version
    impute_accessibility    5min 48s        216295.86 MiB   210960.24 MiB
    normalize_scores        3min 27s        267570.90 MiB   213819.33 MiB
New version
    impute_accessibility    49.6 s          62451.41 MiB    57011.63 MiB
    normalize_scores        2min 16s        160660.14 MiB   106916.97 MiB
Fraction improvement
    impute_accessibility    7x              3.5x            3.7x
    normalize_scores        1.5x            1.6x            2.0x
Results the same
    impute_accessibility    ✅
    normalize_scores        ✅
```
    


# New version

```bash

git checkout optimize_memory_usage

```

## Memory usage impute_accessibility

```python

from diff_features import impute_accessibility
import pickle
from memory_profiler import memory_usage

cistopic_obj = pickle.load(
    open('/staging/leuven/stg_00002/lcb/sdewin/PhD/De_Winter_hNTorg/ATAC_DOWNSTREAM/pycisTopic_w_20221004/outs/cistopic_obj.pkl', 'rb'))

%load_ext memory_profiler
%memit imputed_acc_obj = impute_accessibility(cistopic_obj)

"""
2023-04-28 12:31:44,753 cisTopic     INFO     Imputing region accessibility
2023-04-28 12:31:44,753 cisTopic     INFO     Impute region accessibility for regions 0-20000
2023-04-28 12:31:46,721 cisTopic     INFO     Impute region accessibility for regions 20000-40000
2023-04-28 12:31:48,713 cisTopic     INFO     Impute region accessibility for regions 40000-60000
2023-04-28 12:31:50,623 cisTopic     INFO     Impute region accessibility for regions 60000-80000
2023-04-28 12:31:52,526 cisTopic     INFO     Impute region accessibility for regions 80000-100000
2023-04-28 12:31:54,605 cisTopic     INFO     Impute region accessibility for regions 100000-120000
2023-04-28 12:31:56,689 cisTopic     INFO     Impute region accessibility for regions 120000-140000
2023-04-28 12:31:58,766 cisTopic     INFO     Impute region accessibility for regions 140000-160000
2023-04-28 12:32:01,213 cisTopic     INFO     Impute region accessibility for regions 160000-180000
2023-04-28 12:32:03,270 cisTopic     INFO     Impute region accessibility for regions 180000-200000
2023-04-28 12:32:05,310 cisTopic     INFO     Impute region accessibility for regions 200000-220000
2023-04-28 12:32:07,293 cisTopic     INFO     Impute region accessibility for regions 220000-240000
2023-04-28 12:32:09,441 cisTopic     INFO     Impute region accessibility for regions 240000-260000
2023-04-28 12:32:11,350 cisTopic     INFO     Impute region accessibility for regions 260000-280000
2023-04-28 12:32:13,274 cisTopic     INFO     Impute region accessibility for regions 280000-300000
2023-04-28 12:32:15,381 cisTopic     INFO     Impute region accessibility for regions 300000-320000
2023-04-28 12:32:17,381 cisTopic     INFO     Impute region accessibility for regions 320000-340000
2023-04-28 12:32:19,264 cisTopic     INFO     Impute region accessibility for regions 340000-360000
2023-04-28 12:32:21,260 cisTopic     INFO     Impute region accessibility for regions 360000-380000
2023-04-28 12:32:23,323 cisTopic     INFO     Impute region accessibility for regions 380000-400000
2023-04-28 12:32:25,498 cisTopic     INFO     Impute region accessibility for regions 400000-420000
2023-04-28 12:32:27,529 cisTopic     INFO     Impute region accessibility for regions 420000-440000
2023-04-28 12:32:29,549 cisTopic     INFO     Impute region accessibility for regions 440000-460000
2023-04-28 12:32:31,486 cisTopic     INFO     Impute region accessibility for regions 460000-480000
2023-04-28 12:32:33,430 cisTopic     INFO     Impute region accessibility for regions 480000-500000
2023-04-28 12:32:34,288 cisTopic     INFO     Done!
peak memory: 62451.41 MiB, increment: 57011.63 MiB
"""

```

## Running time impute_accessibility

```python

from diff_features import impute_accessibility
import pickle
from memory_profiler import memory_usage

cistopic_obj = pickle.load(
    open('/staging/leuven/stg_00002/lcb/sdewin/PhD/De_Winter_hNTorg/ATAC_DOWNSTREAM/pycisTopic_w_20221004/outs/cistopic_obj.pkl', 'rb'))

%time imputed_acc_obj = impute_accessibility(cistopic_obj)

"""
2023-04-28 12:33:31,486 cisTopic     INFO     Imputing region accessibility
2023-04-28 12:33:31,486 cisTopic     INFO     Impute region accessibility for regions 0-20000
2023-04-28 12:33:33,411 cisTopic     INFO     Impute region accessibility for regions 20000-40000
2023-04-28 12:33:35,614 cisTopic     INFO     Impute region accessibility for regions 40000-60000
2023-04-28 12:33:37,663 cisTopic     INFO     Impute region accessibility for regions 60000-80000
2023-04-28 12:33:39,796 cisTopic     INFO     Impute region accessibility for regions 80000-100000
2023-04-28 12:33:41,711 cisTopic     INFO     Impute region accessibility for regions 100000-120000
2023-04-28 12:33:43,809 cisTopic     INFO     Impute region accessibility for regions 120000-140000
2023-04-28 12:33:46,063 cisTopic     INFO     Impute region accessibility for regions 140000-160000
2023-04-28 12:33:47,970 cisTopic     INFO     Impute region accessibility for regions 160000-180000
2023-04-28 12:33:50,014 cisTopic     INFO     Impute region accessibility for regions 180000-200000
2023-04-28 12:33:51,962 cisTopic     INFO     Impute region accessibility for regions 200000-220000
2023-04-28 12:33:54,036 cisTopic     INFO     Impute region accessibility for regions 220000-240000
2023-04-28 12:33:55,931 cisTopic     INFO     Impute region accessibility for regions 240000-260000
2023-04-28 12:33:57,981 cisTopic     INFO     Impute region accessibility for regions 260000-280000
2023-04-28 12:34:00,120 cisTopic     INFO     Impute region accessibility for regions 280000-300000
2023-04-28 12:34:02,015 cisTopic     INFO     Impute region accessibility for regions 300000-320000
2023-04-28 12:34:04,003 cisTopic     INFO     Impute region accessibility for regions 320000-340000
2023-04-28 12:34:06,016 cisTopic     INFO     Impute region accessibility for regions 340000-360000
2023-04-28 12:34:08,035 cisTopic     INFO     Impute region accessibility for regions 360000-380000
2023-04-28 12:34:10,130 cisTopic     INFO     Impute region accessibility for regions 380000-400000
2023-04-28 12:34:12,144 cisTopic     INFO     Impute region accessibility for regions 400000-420000
2023-04-28 12:34:14,093 cisTopic     INFO     Impute region accessibility for regions 420000-440000
2023-04-28 12:34:16,019 cisTopic     INFO     Impute region accessibility for regions 440000-460000
2023-04-28 12:34:17,894 cisTopic     INFO     Impute region accessibility for regions 460000-480000
2023-04-28 12:34:19,875 cisTopic     INFO     Impute region accessibility for regions 480000-500000
2023-04-28 12:34:20,726 cisTopic     INFO     Done!
CPU times: user 1min 56s, sys: 2min 33s, total: 4min 30s
Wall time: 49.6 s
"""

pickle.dump(
    imputed_acc_obj,
    open('../tests/imputed_acc_obj_new.pkl', 'wb'))

```

## Memory usage normalize_scores

```python

import pickle
from diff_features import normalize_scores
from memory_profiler import memory_usage

imputed_acc_obj = pickle.load(open('../tests/imputed_acc_obj_new.pkl', 'rb'))

%load_ext memory_profiler
%memit norm_imputed_acc_obj = normalize_scores(imputed_acc_obj)

"""
2023-04-28 13:11:29,090 cisTopic     INFO     Normalizing imputed data
2023-04-28 13:13:45,857 cisTopic     INFO     Done!
peak memory: 160660.14 MiB, increment: 106916.97 MiB
"""

```

## Time usage normalize_scores

```python

import pickle
from diff_features import normalize_scores
from memory_profiler import memory_usage

imputed_acc_obj = pickle.load(open('../tests/imputed_acc_obj_new.pkl', 'rb'))

%time norm_imputed_acc_obj = normalize_scores(imputed_acc_obj)

"""
2023-04-28 13:15:00,593 cisTopic     INFO     Normalizing imputed data
2023-04-28 13:17:17,218 cisTopic     INFO     Done!
CPU times: user 2min 2s, sys: 13.7 s, total: 2min 16s
Wall time: 2min 16s
"""

pickle.dump(
    norm_imputed_acc_obj,
    open('../tests/norm_imputed_acc_obj_new.pkl', 'wb'))

```


# Old version

```bash

git checkout master

```

## Memory usage impute_accessibility

```python

from diff_features import impute_accessibility
import pickle
from memory_profiler import memory_usage

cistopic_obj = pickle.load(
    open('/staging/leuven/stg_00002/lcb/sdewin/PhD/De_Winter_hNTorg/ATAC_DOWNSTREAM/pycisTopic_w_20221004/outs/cistopic_obj.pkl', 'rb'))

%load_ext memory_profiler
%memit imputed_acc_obj = impute_accessibility(cistopic_obj)

"""
2023-04-28 13:26:04,549 cisTopic     INFO     Imputing drop-outs
2023-04-28 13:26:34,636 cisTopic     INFO     Scaling
2023-04-28 13:27:02,759 cisTopic     INFO     Keep non zero rows
2023-04-28 13:27:59,582 cisTopic     INFO     Imputed accessibility sparsity: 0.6921463796367363
2023-04-28 13:27:59,582 cisTopic     INFO     Create CistopicImputedFeatures object
2023-04-28 13:27:59,582 cisTopic     INFO     Making matrix sparse
2023-04-28 13:31:44,037 cisTopic     INFO     Done!
peak memory: 216295.86 MiB, increment: 210960.24 MiB
"""

```

## Running time impute_accessibility

```python

from diff_features import impute_accessibility
import pickle
from memory_profiler import memory_usage

cistopic_obj = pickle.load(
    open('/staging/leuven/stg_00002/lcb/sdewin/PhD/De_Winter_hNTorg/ATAC_DOWNSTREAM/pycisTopic_w_20221004/outs/cistopic_obj.pkl', 'rb'))

%time imputed_acc_obj = impute_accessibility(cistopic_obj)

"""
2023-04-28 13:32:19,567 cisTopic     INFO     Imputing drop-outs
2023-04-28 13:32:49,464 cisTopic     INFO     Scaling
2023-04-28 13:33:19,113 cisTopic     INFO     Keep non zero rows
2023-04-28 13:34:17,823 cisTopic     INFO     Imputed accessibility sparsity: 0.6921463796367363
2023-04-28 13:34:17,824 cisTopic     INFO     Create CistopicImputedFeatures object
2023-04-28 13:34:17,824 cisTopic     INFO     Making matrix sparse
2023-04-28 13:38:07,732 cisTopic     INFO     Done!
CPU times: user 6min 13s, sys: 2min 17s, total: 8min 31s
Wall time: 5min 48s
"""

pickle.dump(
    imputed_acc_obj,
    open('../tests/imputed_acc_obj_old.pkl', 'wb'))

```

## Memory usage normalize_scores

```python

import pickle
from diff_features import normalize_scores
from memory_profiler import memory_usage

imputed_acc_obj = pickle.load(open('../tests/imputed_acc_obj_new.pkl', 'rb'))

%load_ext memory_profiler
%memit norm_imputed_acc_obj = normalize_scores(imputed_acc_obj)

"""
2023-04-28 13:41:49,519 cisTopic     INFO     Normalizing imputed data
2023-04-28 13:45:30,774 cisTopic     INFO     Done!
peak memory: 267570.90 MiB, increment: 213819.33 MiB
"""

```

## Time usage normalize_scores

```python

import pickle
from diff_features import normalize_scores
from memory_profiler import memory_usage

imputed_acc_obj = pickle.load(open('../tests/imputed_acc_obj_new.pkl', 'rb'))

%time norm_imputed_acc_obj = normalize_scores(imputed_acc_obj)

"""
2023-04-28 13:15:00,593 cisTopic     INFO     Normalizing imputed data
2023-04-28 13:17:17,218 cisTopic     INFO     Done!
CPU times: user 2min 2s, sys: 13.7 s, total: 2min 16s
Wall time: 2min 16s
"""

pickle.dump(
    norm_imputed_acc_obj,
    open('../tests/norm_imputed_acc_obj_old.pkl', 'wb'))

```

# Are results the same


```python

import numpy as np
import pickle

imputed_acc_obj_new = pickle.load(open('../tests/imputed_acc_obj_new.pkl', 'rb'))
imputed_acc_obj_old = pickle.load(open('../tests/imputed_acc_obj_old.pkl', 'rb'))

# new code does not remove regions that have 0 value for all cells
# this code was deleted in the new version
"""
log.info("Keep non zero rows")
keep_regions_index = non_zero_rows(imputed_acc)
imputed_acc = imputed_acc[
    keep_regions_index,
]
region_names = subset_list(region_names, keep_regions_index)
"""
imputed_acc_obj_new.subset(features = imputed_acc_obj_old.feature_names)

# the matrix of the new version is already dense, the old one is not
np.allclose(imputed_acc_obj_new.mtx, imputed_acc_obj_old.mtx.todense())
"""
True
"""

del(imputed_acc_obj_new)
del(imputed_acc_obj_old)
gc.collect()

norm_imputed_acc_obj_new = pickle.load(open('../tests/norm_imputed_acc_obj_new.pkl', 'rb'))
norm_imputed_acc_obj_old = pickle.load(open('../tests/norm_imputed_acc_obj_old.pkl', 'rb'))

np.allclose(norm_imputed_acc_obj_new.mtx, norm_imputed_acc_obj_old.mtx)
"""
True
"""

```